pyproject.toml
[project]
name = "crypto-price-tracker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ipykernel>=7.1.0",
    "matplotlib>=3.10.8",
    "numpy>=2.3.5",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]

[dependency-groups]
dev = [
    "black>=25.12.0",
    "pytest>=9.0.2",
]
